+++ spec/options/strategy_spec.rb
@@ -1,20 +1,55 @@
 describe 'a suggestion for a given prefix' do
-  let(:options) { ['_zsh_autosuggest_strategy_default() { suggestion="echo foo" }'] }
+  let(:history_strategy) { '_zsh_autosuggest_strategy_history() { suggestion="history" }' }
+  let(:foobar_strategy) { '_zsh_autosuggest_strategy_foobar() { [[ "foobar baz" = $1* ]] && suggestion="foobar baz" }' }
+  let(:foobaz_strategy) { '_zsh_autosuggest_strategy_foobaz() { [[ "foobaz bar" = $1* ]] && suggestion="foobaz bar" }' }
 
-  it 'is determined by calling the default strategy function' do
-    session.send_string('e')
-    wait_for { session.content }.to eq('echo foo')
+  let(:after_sourcing) do
+    -> do
+      session.run_command(history_strategy)
+    end
   end
 
-  context 'when ZSH_AUTOSUGGEST_STRATEGY is set' do
-    let(:options) { [
-      '_zsh_autosuggest_strategy_custom() { suggestion="echo foo" }',
-      'ZSH_AUTOSUGGEST_STRATEGY=custom'
-    ] }
+  it 'by default is determined by calling the `history` strategy function' do
+    session.send_string('h')
+    wait_for { session.content }.to eq('history')
+  end
 
-    it 'is determined by calling the specified strategy function' do
-      session.send_string('e')
-      wait_for { session.content }.to eq('echo foo')
+  context 'when ZSH_AUTOSUGGEST_STRATEGY is set to an array' do
+    let(:after_sourcing) do
+      -> do
+        session.
+          run_command(foobar_strategy).
+          run_command(foobaz_strategy).
+          run_command('ZSH_AUTOSUGGEST_STRATEGY=(foobar foobaz)')
+      end
+    end
+
+    it 'is determined by the first strategy function to return a suggestion' do
+      session.send_string('foo')
+      wait_for { session.content }.to eq('foobar baz')
+
+      session.send_string('baz')
+      wait_for { session.content }.to eq('foobaz bar')
+    end
+  end
+
+  context 'when ZSH_AUTOSUGGEST_STRATEGY is set to a string' do
+    let(:after_sourcing) do
+      -> do
+        session.
+          run_command(foobar_strategy).
+          run_command(foobaz_strategy).
+          run_command('ZSH_AUTOSUGGEST_STRATEGY="foobar foobaz"')
+      end
+    end
+
+    it 'is determined by the first strategy function to return a suggestion' do
+      session.send_string('foo')
+      wait_for { session.content }.to eq('foobar baz')
+
+      session.send_string('baz')
+      wait_for { session.content }.to eq('foobaz bar')
     end
   end
 end
+